[Refactor] Separate schema into its own file for better modularity
