+++ src/apis/posts/__init__.py
@@ -1,5 +1,6 @@
 from fastapi import APIRouter, status
 
+import src.apis.posts.schema
 from src.apis.posts import create_post, get_post, get_posts
 
 post_router = APIRouter(prefix="/posts", tags=["posts"])
@@ -8,14 +9,14 @@ post_router.add_api_route(
     methods=["POST"],
     path="",
     endpoint=create_post.handler,
-    response_model=create_post.CreatePostResponse,
+    response_model=src.apis.posts.schema.CreatePostResponse,
     status_code=status.HTTP_201_CREATED,
 )
 post_router.add_api_route(
     methods=["GET"],
     path="",
     endpoint=get_posts.handler,
-    response_model=list[get_posts.GetPostResponse],
+    response_model=list[src.apis.posts.schema.GetPostResponse],
     status_code=status.HTTP_200_OK,
 )
 
@@ -23,6 +24,6 @@ post_router.add_api_route(
     methods=["GET"],
     path="/{post_id}",
     endpoint=get_post.handler,
-    response_model=get_post.GetPostResponse,
+    response_model=src.apis.posts.schema.GetPostResponse,
     status_code=status.HTTP_200_OK,
 )
