+++ src/apis/posts/create_post.py
@@ -1,22 +1,11 @@
-import datetime
 from typing import Annotated
 
 from fastapi import Depends
-from pydantic import BaseModel, Field
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from src.apis.dependencies import get_session
+from src.apis.posts.schema import CreatePostRequest, CreatePostResponse
 from src.models.post import Post
-
-
-class CreatePostRequest(BaseModel):
-    contents: str = Field(max_length=500)
-
-
-class CreatePostResponse(BaseModel):
-    id: int
-    contents: str
-    created_at: datetime.datetime
 
 
 async def handler(
