+++ src/apis/posts/get_post.py
@@ -1,18 +1,11 @@
-import datetime
 from typing import Annotated
 
 from fastapi import Depends, HTTPException, status
-from pydantic import BaseModel
 from sqlmodel.ext.asyncio.session import AsyncSession
 
 from src.apis.dependencies import get_session
+from src.apis.posts.schema import GetPostResponse
 from src.models.post import Post
-
-
-class GetPostResponse(BaseModel):
-    id: int
-    contents: str
-    created_at: datetime.datetime
 
 
 async def handler(
